Move ServiceDescriptionController.js to dash controller folder
