+++ test/unit/streaming.controllers.StreamController.js
@@ -18,7 +18,7 @@ import PlaybackControllerMock from './mocks/PlaybackControllerMock';
 import URIFragmentModelMock from './mocks/URIFragmentModelMock';
 import CapabilitiesFilterMock from './mocks/CapabilitiesFilterMock';
 import TextControllerMock from './mocks/TextControllerMock';
-import ServiceDescriptionController from '../../src/streaming/controllers/ServiceDescriptionController';
+import ServiceDescriptionController from '../../src/dash/controllers/ServiceDescriptionController';
 
 const chai = require('chai');
 const spies = require('chai-spies');
